+++ storage/innobase/fts/fts0fts.cc
@@ -338,7 +338,7 @@ fts_get_charset(ulint prtype)
 	}
 #endif /* UNIV_DEBUG */
 
-	uint cs_num = dtype_get_charset_coll(prtype);
+	uint cs_num = (uint) dtype_get_charset_coll(prtype);
 
 	if (CHARSET_INFO* cs = get_charset(cs_num, MYF(MY_WME))) {
 		return(cs);
